+++ extension/src/WorkspaceFunctions.ts
@@ -1,7 +1,7 @@
 import * as fs from "fs";
 import * as vscode from "vscode";
 import * as path from "path";
-import { Settings, Setting } from "./Settings";
+import { Settings, Setting } from "./OldSettings";
 import * as DocumentFunctions from "./DocumentFunctions";
 import { XliffIdToken } from "./ALObject/XliffIdToken";
 import { ALObject } from "./ALObject/ALElementTypes";
